Move tests of training request form to extforms/test directory
